+++ mkovac_zadaca_3/src/main/java/org/foi/uzdiz/mkovac/zadaca_3/strategy/IsporukaNajblizaDostava.java
@@ -94,7 +94,7 @@ public class IsporukaNajblizaDostava implements IsporukaStrategy {
     if (kbrPaketa > najveciKbr)
       kbrPaketa = najveciKbr;
 
-    float postotak = (float) kbrPaketa / (float) najveciKbr;
+    float postotak = (float) (kbrPaketa / najveciKbr);
 
     float gpsLat1 = ulica.getGpsLat1();
     float gpsLon1 = ulica.getGpsLon1();
